scripts/migration/requirements.txt
google-cloud-bigquery==3.24.0


